Corrected size of timer return values.
Correcting functions: get_internal_timer_value(); and
get_timestamp_value(); to indicate 16 bit returned values as opposed to
32 bit.

CAN_TIM & CAN_TIMESTP are 16 bit registers.  Representing them as 32bit
could cause issues when one is attempting to do relative time
calculations and these 16 bit registers have been extended to 32 bits.

+++ due_can.cpp
@@ -506,7 +506,7 @@ uint32_t CANRaw::get_status()
  *
  * \retval The internal CAN free-running timer counter.
  */
-uint32_t CANRaw::get_internal_timer_value()
+uint16_t CANRaw::get_internal_timer_value()
 {
 	return (m_pCan->CAN_TIM);
 }
@@ -517,7 +517,7 @@ uint32_t CANRaw::get_internal_timer_value()
  *
  * \retval The timestamp value.
  */
-uint32_t CANRaw::get_timestamp_value()
+uint16_t CANRaw::get_timestamp_value()
 {
 	return (m_pCan->CAN_TIMESTP);
 }
